requirements.txt
langchain
gradio
langchain-community
openai
atlassian-python-api
langchain_openai
lxml
chromadb
langgraph